requirements.txt
flask_bcrypt
requests
pytest
flask_wtf
flask_sqlalchemy